+++ solvation_analysis/__init__.py
@@ -4,7 +4,7 @@ An MDAnalysis rmodule for solvation analysis.
 """
 
 # Add imports here
-from .solvation import *
+from .selection import *
 
 # Handle versioneer
 from ._version import get_versions
